move code from content-grid.js to header.main.js
