+++ src/js/components/create-markup.js
@@ -1,6 +1,6 @@
 import pageInit from './page-init';
 import ApiService from '../api-service/api-service';
-import { contentCardsRef, contentBtnListRef } from '../content-grid.main';
+import { contentCardsRef, contentBtnListRef } from '../header/header.main';
 
 let badSearchMsg = document.querySelector('.header__warning');
 let body = document.querySelector('BODY');
@@ -8,17 +8,17 @@ const apiService = new ApiService();
 let allWeFoundWarning = document.querySelector('.header__thats-all-msg');
 
 export default function createContentMarkup(element, collection, template) {
-  if (collection.total_results === 0 || collection.length === 0) {
+  if (collection.total_results === 0) {
     badSearchMsg.removeAttribute('style', 'display: none');
     body.addEventListener('click', removeWarning);
     pageInit(apiService, contentCardsRef);
     contentBtnListRef.removeAttribute('style', 'display: none');
-  } else if (collection.total_results < 20 || collection.length < 20) {
+  } else if (collection.total_results < 20) {
     allWeFoundWarning.removeAttribute('style', 'display: none');
     body.addEventListener('click', removeWarning);
   }
 
-  element.insertAdjacentHTML('beforeend', template);  
+  element.insertAdjacentHTML('beforeend', template);
 }
 
 function removeWarning() {
