+++ src/js/components/on-search.js
@@ -1,4 +1,3 @@
-import getGenres from './get-genres';
 import renderContent from './render-content';
 
 export default async function pageOnSearch(api, elemtRef) {
