+++ src/js/main.js
@@ -1,3 +1,2 @@
 import './header/header.main';
-import './content-grid.main';
 import './components/goup-btn';
